backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
asyncpg==0.29.0
httpx==0.27.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
python-dotenv==1.0.1
pydantic[email]==2.7.1
apscheduler==3.10.4
pandas==2.2.2
openpyxl==3.1.2
aiosmtplib==3.0.1
email-validator==2.1.1
